Change all Int64 to either Int or Double

+++ EasyWeather/Model/Current/CurrentResponse.swift
@@ -31,7 +31,7 @@ class CurrentResponse: Mappable {
     var clouds: CurrentClouds?
     var dateAsDouble: Double?
     var sys: CurrentSys?
-    var id: Int64?
+    var id: Double?
     var name: String?
     
     required init?(map: Map) {
